ml/model_requirements.txt
S3Fs==2022.11.0
patool==1.12
pandas==1.3.5
pydantic==1.10.2
Pillow==9.3.0
numpy==1.21.6
tensorflow==2.11.0
keras
boto3==1.24.59
unrar==0.4
matplotlib